Source Map support
This commit adds support for source maps. Source maps are enabled by
setting the option:

```js
{
	sourceMaps: true
}
```

Optionally the original source code can be included with the source maps
(this is useful in a production environment where the original sources
 are not accessible) by setting the `sourceMapsContent` option to true:

```js
{
	sourceMaps: true,
	sourceMapsContent: true
}
```

To make this feature work I had to change the signature for
`transpile.to` to return an object. The object will be like:

```js
{
	code: "", // Source code
	map: SourceMapConsumer, // from the source-maps project
	ast: {} // A Mozilla Parser API compatible AST (from esprima)
}
```

Fixes #23