Replace all "module" (and other cjs) identifiers
transpile would only replace it if it was part of a member expression,
this time it would replace all identifiers but those that are on the
right side of a member expression. E.g:

    "module.exports = {};"

should be transformed to:

    "stealModule.exports = {};"

Note that the "exports" in the right side of the member expression
should not be converted to "stealExports".